android: compileSdkVersion to 34 (Android 14)
As we said on the last one of these, fc2dab717:

  This setting is not to be confused with the targetSdkVersion. The
  latter goes into the built manifest, and affects a wide range of
  behavior, so bumping it requires careful testing. This is used
  purely at build time, and should have no effect on runtime
  behavior.

This value needs to be at least as high as targetSdkVersion, and we
need to bump that to 34 soon (#5877). So, do this to prepare for
that.

The build warnings change slightly, but the new warnings all seem
acceptable.  Details and discussion starting at:
  https://github.com/zulip/zulip-mobile/pull/5879#issuecomment-2362286748

Related: #5877