deps: Use a patched rn-fetch-blob to fix Android 14+ crash
This updates rn-fetch-blob to a version with one cherry-picked commit
atop the v0.11.2 we were previously using:
  https://github.com/zulip/rn-fetch-blob/commit/e17b2a51c

The rn-fetch-blob package itself is no longer maintained:
  https://github.com/joltup/rn-fetch-blob/issues/863

There's a fork that is maintained, `react-native-blob-util`:
  https://npmjs.com/package/react-native-blob-util

That fork addressed the Android 14 requirement to set an
explicit RECEIVER_EXPORTED flag (#5903), merging a fix:
  https://github.com/RonRadtke/react-native-blob-util/issues/303
  https://github.com/RonRadtke/react-native-blob-util/pull/306

Our fix is cherry-picked from that one.

(If this app weren't a legacy codebase in maintenance mode, we'd take
the time to upgrade this dependency more generally, probably switching
to the react-native-blob-util fork.  As is, we'll skip that and use
the time to focus on completing the new Flutter app.)

Fixes: #5903