api: Don't allow connecting to servers <5.0; show nag banner on <7.0
See 8552afce9 for the previous one of these.

The Zulip Server 5 release went out over 3 years ago, so the 4.x
releases are all long out of our announced 18-month compatibility
window:
  https://blog.zulip.com/2022/03/29/zulip-5-0-released/
  https://zulip.readthedocs.io/en/latest/overview/release-lifecycle.html#client-apps

So we can cheerfully refuse to connect to 4.x releases; our "Server
not supported" nag banner has been asking server admins on those
versions to upgrade since 8552afce9, which the changelog says went
out in v27.216, dated 2023-10-10.

The last 6.x release, 6.2, went out on 2023-05-19. That's also
outside the 18-month window, so it's time to start showing the nag
banner on 5.x and 6.x.

The last 7.x release, 7.5, went out on 2023-11-16, which is a month
and a bit shy of 18 months, so it's not yet time to show the banner
for versions 7.x.