api: Don't allow connecting to servers <7.0; show nag banner on <8.0
The last one of these was 48f1e4724 on 2025-04-04, released in
v27.234 on 2025-04-24.

As mentioned there, the last 6.x release, 6.2, went out on
2023-05-19, which is outside the 18-month window. We've been showing
the nag banner for that release since v27.234, so we can expect
server admins to have seen and acted on it, and we can refuse to
connect.

The last 7.x release, 7.5, went out on 2023-11-16, which turns 18
months old tomorrow, 2025-05-16. So we can start to show the nag
banner for that.